+++ changelog.md
@@ -5,6 +5,8 @@
 - Redesigned tabs in charts
 - Deployed to new domain ebcharts.com
 - Added current year estimate in years graph
+- Added missing hotelpartner: Smart Hotel
+- Added missing resturant: Folk & Fe
 
 ### 08.04.2025